Tests/LibWeb/Text/expected/navigation/iframe-referrer-policy.txt

no-referrer: PASS
origin: PASS
unsafe-url: PASS
